+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Removed the processing of trailer metadata on stream calls 
+
 ## v3.26.7
 * Updated the `ydb-go-genproto` dependency